positional wip
Tried to convert positional projection to $elemMatch, but had difficulty with finding the "sub " match expression to put into the $elemMatch and serializing it correctly,
because the match expression has field names attached to it. E.g. If the query is on x.a: 1 and the positional projection is x.$, we'd want an elemMatch: {a: 1}. If you just walk the match
expression tree to find the children of 'x' and then serialize, you'll still get "x.a": whatever in the elemMatch.

This was hard to explain.

Tomorrow I'm gonna try adding $_positionalInternalThing and just forward the entire top-level match expression over. There are two ways to do this:
1) Reserialize the match expr into the $positionalInternalThing
2) Pass the match expression into parsed_agg_projection and then have $positionalInternalThing use it