almost have it compiling again, trying to add UniqueBuilder
We have to switch to CRTP because some methods return *this but other methods
only make sense on the BSONObjBuilder which uses shared buffer. For example the obj() method
calls shareOwnerShipWith(), and assumes the underlying buf builder is shared. So obj() can only
be a method on BSONObjBuilder and not BSONObjBuilderBase. If I call a chain like

BSONObjBuilder{}.appendX().appendY().obj()

appendX() and appendY() will return a BSONObjBuilderBase<BufBuilder> NOT BSONObjBuilder(), so we can't call obj(). So it's a compile error.

So instead I'm using CRTP.